qlibrary update: base->core and a plural
