+++ qiskit_metal/qlibrary/core/__init__.py
@@ -13,7 +13,7 @@
 # that they have been altered from the originals.
 
 from ...toolbox_python.attr_dict import Dict
-from .. import is_component
+from ... import is_component
 from .base import QComponent
 from .qroute import QRoute, QRoutePoint, QRouteLead
 from .qubit import BaseQubit
