+++ qiskit_metal/qlibrary/couplers/tee.py
@@ -11,48 +11,42 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
+""""""
 
 from qiskit_metal import draw, Dict
-from qiskit_metal.qlibrary.base import QComponent
+from qiskit_metal.qlibrary.core import QComponent
 import numpy as np
 
 
-class CPWHangerT(QComponent):
+class CPWT(QComponent):
     """Generates a three pin (+) structure comprised of a primary two pin CPW
     transmission line, and a secondary one pin neighboring CPW transmission
-    line that is capacitively/inductively coupled to the primary. Such a
-    structure can be used, as an example, for generating CPW resonator hangars
-    off of a transmission line.
+    line that is capacitively coupled to the primary. Such a structure can be
+    used, as an example, for generating CPW resonator hangars off of a
+    transmission line. (0,0) represents the center position of the component.
 
     Inherits QComponent class.
 
     ::
 
-        +----------------------------+
-        ------------------------------
-        |
-        |
-        |
-        |
-        +
+                  (0,0)
+        +--------------------------+
+                    |
+                    |
+                    |
+                    |
+                    |
+                    +
 
-    Default Options:
+    Options:
         * prime_width: '10um' -- The width of the trace of the two pin CPW transmission line
         * prime_gap: '6um' -- The dielectric gap of the two pin CPW transmission line
         * second_width: '10um' -- The width of the trace of the one pin CPW transmission line
-        * second_gap: '6um' -- The dielectric gap of the one pin CPW transmission line
-        * coupling_space: '3um' -- The amount of ground plane between the two transmission lines
-        * coupling_length: '100um' -- The length of parallel between the two transmission lines
-          note: this includes the distance of the curved second of the second line
-        * down_length: '100um' -- The length of the hanging part of the resonator, including the curved region
-        * fillet: '25um'
-        * pos_x: '0um' -- The x position of the ground termination.
-        * pos_y: '0um' -- The y position of the ground termination.
-        * rotation: '0' -- The direction of the termination. 0 degrees is +x, following a
+        * second_gap: '6um' -- The dielectric gap of the one pin CPW transmission line (also for the capacitor)
+        * t_length: '50um' -- The length for the t branches
+        * pos_x/_y: '0um' -- The x/y position of the center of the primary transmission line.
+        * orientation: '0' -- The direction of the primary transmission line. 0 degrees is +x, following a
           counter-clockwise rotation (eg. 90 is +y)
-        * mirror: False -- Flips the hanger around the y-axis
-        * open_termination: True -- sets if the termination of the second line at the coupling side
-          is an open to ground or short to ground
         * chip: 'main' -- The chip the pin should be on.
         * layer: '1' -- Layer the pin is on. Does not have any practical impact to the short.
     """
@@ -65,15 +59,10 @@ class CPWHangerT(QComponent):
                            prime_gap='6um',
                            second_width='10um',
                            second_gap='6um',
-                           coupling_space='3um',
-                           coupling_length='100um',
-                           down_length='100um',
-                           fillet='25um',
+                           t_length='50um',
                            pos_x='0um',
                            pos_y='0um',
-                           rotation='0',
-                           mirror=False,
-                           open_termination=True,
+                           orientation='0',
                            chip='main',
                            layer='1')
     """Default connector options"""
@@ -81,60 +70,37 @@ class CPWHangerT(QComponent):
     def make(self):
         """Build the component."""
         p = self.p
-
-        prime_cpw_length = p.coupling_length * 2
-        second_flip = 1
-        if p.mirror:
-            second_flip = -1
+        prime_cpw_length = p.t_length * 2
 
         #Primary CPW
         prime_cpw = draw.LineString([[-prime_cpw_length / 2, 0],
                                      [prime_cpw_length / 2, 0]])
 
         #Secondary CPW
-        second_down_length = p.down_length
-        second_y = -p.prime_width / 2 - p.prime_gap - p.coupling_space - p.second_gap - p.second_width / 2
-        second_cpw = draw.LineString(
-            [[second_flip * (-p.coupling_length / 2), second_y],
-             [second_flip * (p.coupling_length / 2), second_y],
-             [
-                 second_flip * (p.coupling_length / 2),
-                 second_y - second_down_length
-             ]])
-
-        second_termination = 0
-        if p.open_termination:
-            second_termination = p.second_gap
-
-        second_cpw_etch = draw.LineString(
-            [[
-                second_flip * (-p.coupling_length / 2 - second_termination),
-                second_y
-            ], [second_flip * (p.coupling_length / 2), second_y],
-             [
-                 second_flip * (p.coupling_length / 2),
-                 second_y - second_down_length
-             ]])
+        second_cpw = draw.LineString([[0, -p.prime_width / 2], [0,
+                                                                -p.t_length]])
 
         #Rotate and Translate
-        c_items = [prime_cpw, second_cpw, second_cpw_etch]
-        c_items = draw.rotate(c_items, p.rotation, origin=(0, 0))
+        c_items = [prime_cpw, second_cpw]
+        c_items = draw.rotate(c_items, p.orientation, origin=(0, 0))
         c_items = draw.translate(c_items, p.pos_x, p.pos_y)
-        [prime_cpw, second_cpw, second_cpw_etch] = c_items
+        [prime_cpw, second_cpw] = c_items
 
         #Add to qgeometry tables
         self.add_qgeometry('path', {'prime_cpw': prime_cpw},
-                           width=p.prime_width)
+                           width=p.prime_width,
+                           layer=p.layer)
         self.add_qgeometry('path', {'prime_cpw_sub': prime_cpw},
                            width=p.prime_width + 2 * p.prime_gap,
-                           subtract=True)
+                           subtract=True,
+                           layer=p.layer)
         self.add_qgeometry('path', {'second_cpw': second_cpw},
                            width=p.second_width,
-                           fillet=p.fillet)
-        self.add_qgeometry('path', {'second_cpw_sub': second_cpw_etch},
+                           layer=p.layer)
+        self.add_qgeometry('path', {'second_cpw_sub': second_cpw},
                            width=p.second_width + 2 * p.second_gap,
                            subtract=True,
-                           fillet=p.fillet)
+                           layer=p.layer)
 
         #Add pins
         prime_pin_list = prime_cpw.coords
@@ -149,6 +115,6 @@ class CPWHangerT(QComponent):
                      width=p.prime_width,
                      input_as_norm=True)
         self.add_pin('second_end',
-                     points=np.array(second_pin_list[1:]),
+                     points=np.array(second_pin_list),
                      width=p.second_width,
                      input_as_norm=True)
