+++ qiskit_metal/qlibrary/couplers/tunable_coupler_01.py
@@ -14,7 +14,7 @@
 
 import numpy as np
 from qiskit_metal import draw, Dict
-from qiskit_metal.qlibrary.base.qubit import BaseQubit
+from qiskit_metal.qlibrary.core import BaseQubit
 
 
 class TunableCoupler01(BaseQubit):
