+++ qiskit_metal/qlibrary/lumped/cap_3_interdigital.py
@@ -14,7 +14,7 @@
 
 from qiskit_metal import draw
 from qiskit_metal.toolbox_python.attr_dict import Dict
-from qiskit_metal.qlibrary.base.base import QComponent
+from qiskit_metal.qlibrary.core import QComponent
 
 
 class CapThreeFingers(QComponent):
