+++ qiskit_metal/qlibrary/qubits/transmon_concentric.py
@@ -15,7 +15,7 @@
 import numpy as np
 from math import *
 from qiskit_metal import draw, Dict
-from qiskit_metal.qlibrary.base.qubit import BaseQubit
+from qiskit_metal.qlibrary.core import BaseQubit
 
 
 class TransmonConcentric(BaseQubit):
