+++ qiskit_metal/qlibrary/qubits/transmon_cross.py
@@ -14,7 +14,7 @@
 
 import numpy as np
 from qiskit_metal import draw, Dict
-from qiskit_metal.qlibrary.base.qubit import BaseQubit
+from qiskit_metal.qlibrary.core import BaseQubit
 
 
 class TransmonCross(BaseQubit):  # pylint: disable=invalid-name
