+++ qiskit_metal/qlibrary/sample_shapes/rectangle_hollow.py
@@ -13,7 +13,7 @@
 # that they have been altered from the originals.
 
 from qiskit_metal import draw, Dict
-from qiskit_metal.qlibrary.base import QComponent
+from qiskit_metal.qlibrary.core import QComponent
 
 
 class RectangleHollow(QComponent):
