+++ qiskit_metal/tests/test_qlibrary_1_instantiate.py
@@ -21,11 +21,11 @@
 import unittest
 import numpy as np
 
-from qiskit_metal.qlibrary.base.base import QComponent
-from qiskit_metal.qlibrary.base.qroute import QRoute
-from qiskit_metal.qlibrary.base.qroute import QRouteLead
-from qiskit_metal.qlibrary.base.qroute import QRoutePoint
-from qiskit_metal.qlibrary.base.qubit import BaseQubit
+from qiskit_metal.qlibrary.core import QComponent
+from qiskit_metal.qlibrary.core import QRoute
+from qiskit_metal.qlibrary.core import QRouteLead
+from qiskit_metal.qlibrary.core import QRoutePoint
+from qiskit_metal.qlibrary.core import BaseQubit
 from qiskit_metal.qlibrary.sample_shapes.circle_caterpillar import CircleCaterpillar
 from qiskit_metal.qlibrary.sample_shapes.circle_raster import CircleRaster
 from qiskit_metal.qlibrary.sample_shapes.rectangle import Rectangle
@@ -33,9 +33,9 @@ from qiskit_metal.qlibrary.sample_shapes.rectangle_hollow import RectangleHollow
 from qiskit_metal.qlibrary.sample_shapes.n_gon import NGon
 from qiskit_metal.qlibrary.sample_shapes.n_square_spiral import NSquareSpiral
 from qiskit_metal.qlibrary.lumped.cap_n_interdigital import CPWFingerCap
-from qiskit_metal.qlibrary.coupler.coupled_line_tee import CPWHangerT
-from qiskit_metal.qlibrary.coupler.cap_n_interdigital_tee import CPWTFingerCap
-from qiskit_metal.qlibrary.coupler.tee import CPWT
+from qiskit_metal.qlibrary.couplers.coupled_line_tee import CPWHangerT
+from qiskit_metal.qlibrary.couplers.cap_n_interdigital_tee import CPWTFingerCap
+from qiskit_metal.qlibrary.couplers.tee import CPWT
 from qiskit_metal.qlibrary.terminations.open_to_ground import OpenToGround
 from qiskit_metal.qlibrary.terminations.short_to_ground import ShortToGround
 from qiskit_metal.qlibrary.tlines.anchored_path import RouteAnchors
@@ -53,7 +53,7 @@ from qiskit_metal.qlibrary.qubits.transmon_cross_fl import TransmonCrossFL
 from qiskit_metal.qlibrary.qubits.transmon_pocket import TransmonPocket
 from qiskit_metal.qlibrary.qubits.transmon_pocket_cl import TransmonPocketCL
 from qiskit_metal.qlibrary.qubits.transmon_pocket_6 import TransmonPocket6
-from qiskit_metal.qlibrary.coupler.tunable_coupler_01 import TunableCoupler01
+from qiskit_metal.qlibrary.couplers.tunable_coupler_01 import TunableCoupler01
 from qiskit_metal import designs
 from qiskit_metal.qlibrary._template import MyQComponent
 from qiskit_metal.tests.assertions import AssertionsMixin
